DataHub: enable FE to DataHub, implement dataHubProvider
