+++ forward_engineering/ddlProvider.js
@@ -56,6 +56,7 @@ const {
 	prepareObjectTagsData,
 	isEmptyTags,
 } = require('./helpers/tagHelper');
+const { createView, hydrateView, hydrateViewColumn } = require('./helpers/viewHelper');
 
 const DEFAULT_SNOWFLAKE_SEQUENCE_START = 1;
 const DEFAULT_SNOWFLAKE_SEQUENCE_INCREMENT = 1;
@@ -582,76 +583,12 @@ module.exports = (baseProvider, options, app) => {
 		},
 
 		createView(viewData, dbData, isActivated) {
-			const orReplace = preSpace(viewData.orReplace && 'OR REPLACE');
-			const ifNotExist = preSpace(viewData.ifNotExist && 'IF NOT EXISTS');
-			const { columnList, tableColumns, tables } = viewData.keys.reduce(
-				(result, key) => {
-					result.columnList.push({
-						name: `${getName(viewData.isCaseSensitive, key.alias || key.name)}`,
-						isActivated: key.isActivated,
-						comment: preSpace(
-							key.definition?.description &&
-								`COMMENT ${escapeString(scriptFormat, key.definition.description)}`,
-						),
-					});
-					result.tableColumns.push({
-						name: `${getName(viewData.isCaseSensitive, key.entityName)}.${getName(viewData.isCaseSensitive, key.name)}`,
-						isActivated: key.isActivated,
-					});
-
-					if (key.entityName) {
-						const tableName = getFullName(key.dbName, key.entityName);
-
-						if (!result.tables.includes(tableName)) {
-							result.tables.push(tableName);
-						}
-					}
-
-					return result;
-				},
-				{
-					columnList: [],
-					tableColumns: [],
-					tables: [],
-				},
-			);
-
-			if (isEmpty(tables) && !viewData.selectStatement) {
-				return '';
-			}
-
-			const viewColumns = viewColumnsToString(tableColumns, isActivated);
-			const selectStatement = getViewSelectStatement({
-				tables,
+			return createView({
 				viewData,
-				viewColumns,
-			});
-
-			const tagStatement = getTagStatement({
-				tags: viewData.viewTags,
-				isCaseSensitive: viewData.isCaseSensitive,
-				indent: '',
-			});
-
-			const clustering = viewData.materialized
-				? keyHelper.getClusteringKey({
-						clusteringKey: viewData.clusteringKey,
-						isParentActivated: isActivated,
-					})
-				: undefined;
-
-			return assignTemplates(templates.createView, {
-				orReplace,
-				ifNotExist,
-				secure: preSpace(viewData.secure && 'SECURE'),
-				materialized: preSpace(viewData.materialized && 'MATERIALIZED'),
-				name: viewData.fullName,
-				column_list: viewColumnsToString(columnList, isActivated),
-				copy_grants: viewData.copyGrants ? 'COPY GRANTS\n' : '',
-				comment: viewData.comment ? `COMMENT=${escapeString(scriptFormat, viewData.comment)}\n` : '',
-				select_statement: selectStatement,
-				tag: tagStatement ? tagStatement + '\n' : '',
-				clustering,
+				isActivated,
+				scriptFormat,
+				getViewSelectStatement,
+				keyHelper,
 			});
 		},
 
@@ -1006,39 +943,11 @@ module.exports = (baseProvider, options, app) => {
 		},
 
 		hydrateView({ viewData, entityData }) {
-			const firstTab = entityData[0];
-			const { databaseName, schemaName } = viewData.schemaData;
-			const viewName = getName(firstTab.isCaseSensitive, viewData.name);
-			const fullName = getFullName(getFullName(databaseName, schemaName), viewName);
-
-			return {
-				...viewData,
-				orReplace: firstTab.orReplace,
-				ifNotExist: firstTab.ifNotExist,
-				name: getName(firstTab.isCaseSensitive, viewData.name),
-				selectStatement: firstTab.selectStatement,
-				isCaseSensitive: firstTab.isCaseSensitive,
-				copyGrants: firstTab.copyGrants,
-				comment: firstTab.description,
-				secure: firstTab.secure,
-				materialized: firstTab.materialized,
-				fullName,
-				clusteringKey: firstTab.clusteringKey,
-				viewTags: firstTab.viewTags ?? [],
-			};
+			return hydrateView({ viewData, entityData });
 		},
 
 		hydrateViewColumn(data) {
-			if (!data.entityName) {
-				return data;
-			}
-
-			return {
-				...data,
-				name: getName(data.definition?.isCaseSensitive, data.name),
-				dbName: getName(head(data.containerData)?.isCaseSensitive, data.dbName),
-				entityName: getName(head(data.entityData)?.isCaseSensitive, data.entityName),
-			};
+			return hydrateViewColumn(data);
 		},
 
 		commentIfDeactivated(statement, data, isPartOfLine) {
